requirements.txt
# ============================================================
# SPECTROLITE — Degenerate Gambit v2.0
# requirements.txt
# ============================================================

# ── Blockchain ───────────────────────────────────────────────
web3==6.15.1
solana==0.30.2
anchorpy==0.19.1
ccxt==4.2.15
solders==0.19.0

# ── Machine Learning / AI ────────────────────────────────────
torch==2.2.0
torchvision==0.17.0
transformers==4.38.0
scikit-learn==1.4.0
sentence-transformers==2.6.1
opencv-python==4.9.0.80
open-clip-torch==2.24.0
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0

# ── Orchestration ────────────────────────────────────────────
crewai==0.28.0
langchain==0.1.12
langchain-openai==0.0.8
langgraph==0.0.38
openai==1.13.3

# ── Automation ───────────────────────────────────────────────
playwright==1.42.0
pyotp==2.9.0
aiohttp==3.9.3
httpx==0.27.0
websockets==12.0
asyncio==3.4.3

# ── Visualization / Dashboard ────────────────────────────────
streamlit==1.32.0
plotly==5.19.0
pyttsx3==2.90
pillow==10.2.0

# ── Security ─────────────────────────────────────────────────
cryptography==42.0.5
python-dotenv==1.0.1
pynacl==1.5.0

# ── Infrastructure ───────────────────────────────────────────
redis==5.0.3
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic-settings==2.2.1

# ── Monitoring ───────────────────────────────────────────────
prometheus-client==0.20.0
structlog==24.1.0

# ── Notifications ────────────────────────────────────────────
python-telegram-bot==21.0.1
reportlab==4.1.0   # PDF generation for autopsy report
jinja2==3.1.3

# ── Testing ──────────────────────────────────────────────────
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-mock==3.12.0
hypothesis==6.99.4

# ── Utilities ────────────────────────────────────────────────
pyyaml==6.0.1
click==8.1.7
rich==13.7.1
tenacity==8.2.3
apify-client==1.6.4
